refactor(git-diff): name the long-line module for its job
budget.js was named for a category rather than a job, unlike every other
module in the plugin, and it held one constant with no code beside it.

Rename it to longlines.js, which is what it does: replace a file whose
longest line is over the ceiling with git's binary placeholder, before the
patch is parsed. Move MAX_PATCH_BYTES to payload.js, beside the other
render-cost ceilings and next to tooBigVerdict, the function that reads it.
It stays an injected parameter there, so the payload tests can drive it
with small numbers.

No behaviour change. All five plugin tests pass.
